Documentation and Examples (#14)
* Update code

* Update flow.json

* Start api documentation

* Add example tests

* Remove spaces in getArrayType result

* Add method to completely remove all spaces in a string

* Export additional methods and arguments

* Add more tests and API documentation

* Replace default dependency

* Update templates

* Add fix for non existent folders

* Add test Cadence templates for generator

* Add generator tests

* Regenerate templates

* Add documentation for Prettier and generator related methods

* Reformat headers. refactor method description

* Update environment logic. add setEnvironment method

* Add env tests

* Export additional methods

* Add docs for environment methods

* Ad aliases

* Add transaction documentation

* Fix dev tests

* Refactor for seal status

* Update handlebars templates to default interaction props to empty object

+++ src/processor.js
@@ -40,7 +40,9 @@ const getFetchUrl = (input) => {
 
 const TEMP_REPO_FOLDER = path.resolve(process.cwd(), "./temp-generator-repo");
 const clean = () => {
-  fs.rmdirSync(TEMP_REPO_FOLDER, { recursive: true });
+  if (fs.existsSync(TEMP_REPO_FOLDER)) {
+    fs.rmdirSync(TEMP_REPO_FOLDER, { recursive: true });
+  }
 };
 
 export const getBranchesList = (branches, remotes) => {
@@ -138,7 +140,7 @@ export const processFolder = async (input, output, options = {}) => {
     const pathPackages = packages.slice(0, -1);
     const file = packages.slice(-1)[0];
 
-    const ixDependency = options.dependency || "flow-js-testing";
+    const ixDependency = options.dependency || "flow-cadut";
 
     const code = readFile(path).replace(/`/g, "'");
     const name = underscoreToCamelCase(file.replace(".cdc", ""));
